Complete moving hoplite to a class.
It wasn't just as simple as adding class hoplite(): at the top and patting
myself on the back, apparently.

Also moved reading weight data and rendering to their own functions.  This
will help when web interface comes; that script will become responsible for
updating weight data and sending results to the LCD regularly.