Implement IPC via shared memory.
hoplite now creates a shared memory file and a semaphore via posix_ipc.

The shared memory file contains the full config and current data for keg
weights.  Eventually hoplite will read back any changes to the config so the
webserver can update that via a more friendly interface.

The semaphore is for locking the shmem file, to prevent conflicts.