Add changing keg sizes via web interface.
hoplite-web.py: Add dropdowns for keg sizes, populated automatically from the
local Hoplite instance, and add the necessary bits to update the shared config
appropriately when said sizes are changed.

hoplite.py: Add size_name to config, to store the name of the current size of
the keg (we were already storing the raw size data).  This should allow for
overriding built-in size data later via a 'custom' name.