Rework CO2 handling slightly
Added a hack to allow specifying which channel of the CO2 module
has the CO2 platform attached.  In the future, I'll add a flag
to all of the HX711 channels to allow specifying that channel as
a CO2 cylinder, and also handle multiple CO2 cylinders.

I also modified debug_mgs() to add the name of the calling function,
which should make debug easier.