Investigate issues with adding a keg
For some reason it takes several seconds to get an actual selection
from the keg size drop-down.  It'll be on the label, but the internals
of Remi have the value as None.

I need to look into the performance issues since I switched to
Python 3...  I also need to find a way to gray out the OK button
if mandatory data fields aren't populated.